requirements.txt
fastapi
python-dotenv
python-dotenv
pydantic
pydantic-settings
psycopg2-binary
sqlalchemy
python-multipart
loguru
pandas
numpy
langchain
langchain-openai

langchain-community
langchain-core
uvicorn
openpyxl
langsmith